Make the single function (minAngDist) part of scorecard rather than in it's own file
